% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKconverterApp.R
\name{PKconverterApp}
\alias{PKconverterApp}
\title{Shiny App for converting the pharmacokinetic parameters}
\usage{
PKconverterApp()
}
\description{
Pharmacokinetics is the study of drug absorption, distribution, metabolism,
and excretion. The pharmacokinetics model explains that how the drug
concentration change as the drug moves through the different compartments
of the body. For pharmacokinetic modeling and analysis, it is essential
to understand the basic pharmacokinetic parameters. All parameters are
considered, but only some of parameters are used in the model.
Therefore,we need to convert the estimated parameters to the other
parameters after fitting the specific pharmacokinetic model.
For Shiny App on the web, visit
\href{https://ek-lee.shinyapps.io/PKconverter/}{PKconverter shiny app},
}
\examples{
#PKconverterApp()
}
